Sorting sidebar links
Another change from vanilla.  Nothing is changed here except the order of the sidebar links.  The important bit is moving game options so it is always at the end of the list and seperated by whitespace.

After that, the goal was to group similar things.  I picked two categories for this in the vanilla commit, "visiting real places" and "setting rules from your desk."  They don't need to be labeled, it's just implied.

So I grouped the very similar policies and edicts, and put them with universal rules.

Then moved the PR, sec, mil, and rebs up in the list to be with the other "places."  Should be more obvious at a glance where things are.