induced NCS -> quick list
Adds a game option to enable a jQuery in page scrolling link box for
navigation to specific slaves (by name) in slave lists.

Adds backwards compatibility for our option variable.

Update the slave Summary to, if enabled, build a quick slave in-page
jQuery scrollto navigation for getting to individual slaves quickly.

While this may be overkill for facitlities with 1 to 5 slaves, more
than that it becomes very useful.

-	modified:   src/uncategorized/BackwardsCompatibility.tw
-	modified:   src/uncategorized/options.tw
-	modified:   src/uncategorized/slaveSummary.tw