induced NCS, Quicklist penthouse tabs
Had a rethink and just put some filters on the tabbed content, so that
it only activates on the 'All' tab (which is the default tab).  The idea
is if you're clicking on the other tabs, you've already narrowed your
selection, but in the 'All' tab, the quick list will appear and work as
expected.

-	new file:   src/js/hTagMacroJS.tw
-	modified:   src/uncategorized/slaveSummary.tw