inducedncs-QL-devtr-color click button sort
A lot of background work was needed to make this happen.  Had to pass
the devotion and trust as data-foo attributes to the buttons so we can
build a sort based on the values.  So updated the hTag macro to work
better (correctly).
-	modified:   src/js/hTagMacroJS.tw

Added new CSS file to include our old/modified css and our new css in:
-	new file:   src/gui/css/quickListCSS.tw

Added new JS file to include our quick list specific functions in:
-	new file:   src/js/quickListJS.tw

Updated JS file to hold some more globally useful JS functions:
-	modified:   src/js/utilJS.tw

Needed to set up a means to hold the classes for coloration for both
types of coloration, the devotion theming and the trust theming; Used a
wrapper div that either holds a Devotion or Trust class, then when we
toggle that class we can fall into the trust or devotion pallet.

In the process, built a JS function to encapsulate the Name building
that's currently in the twine/sugarcube, rewritten as javascript for
simplicity, speed and portability.

Built a one use global variable to hold either devotion sort or trust
sort, and instantly defaulted to devotion, so not needing a backwards
compatibility update.

-	modified:   src/uncategorized/slaveSummary.tw